fix: popover not showing when using new architecture
`NativeModules.UIManager.measure(...)` does not work when new architecture is enabled.
`ref.current.measureInWindow(...)` works on both old and new architecture and returns the same values, except one case: when the status bar is translucent on Android, it substracts the height of the status bar from the "y" value (see https://github.com/SteffeyDev/react-native-popover-view/pull/182#issuecomment-2537382284). This means we need to adjust the verticalOffset on Android when the status bar is translucent. To ensure that we always know whether the status bar is translucent or not, we'll default the value of statusBarTranslucent to true when not passed explicitly (which is more similar to iOS) instead of passing undefined to RN Modal.
Since the only intended purpose the `verticalOffset` prop was to correct manually the position on Android and now it's adjusted internally, we don't need that prop anymore and we are removing it.